AppPagination moved to FindOfers component
